requirements.txt
requests==2.26.0
flask==2.0.2
python-dotenv==0.19.2